+++ service/src/ec_control/mod.rs
@@ -4,6 +4,12 @@
 pub mod ec_manager;
 mod read;
 mod write;
+use std::io::{Read, Seek, Write};
+
+type RcWrapper<T> = std::rc::Rc<std::cell::RefCell<T>>;
+
+pub(crate) trait RW: Read + Write + Seek + std::fmt::Debug {}
+impl<T: Read + Write + Seek + std::fmt::Debug> RW for T {}
 
 // For the cleaner.
 pub(crate) use write::ECWriter;