+++ service/CHANGELOG.md
@@ -5,6 +5,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Changed
+
+- Refactoring code:
+  - `Arc` have been replaced by `Rc`, since the code is single-threaded now.
+  - `Mutex` and `RwLock` have been replaced by `RefCell` where appropriated.
+  - Remove `Unpin` trait.
+
 ## [v0.2.0] - 2020-09-25
 
 ### Added
@@ -16,3 +23,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 Initial version.
 
 [v0.2.0]: https://github.com/musikid/fancy/compare/fancy-service-0.1.0...fancy-service-0.2.0
+[v0.1.0]: https://github.com/musikid/fancy/compare/fancy-service-0.1.0